Move USN code to the security package
